backend/requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
pydantic==2.10.6
python-dotenv==1.0.1
openai==1.63.2
numpy==2.2.3
sqlalchemy==2.0.38
psycopg[binary]==3.2.5
alembic==1.14.1
pytest==8.3.4
